Remove to_host transfer option

+++ utoolbox/transform/rotate.cu
@@ -1,5 +1,3 @@
-#include <pycuda-helpers.hpp>
-
 texture<float, cudaTextureType2D, cudaReadModeElementType> rot2_tex;
 
 __global__
@@ -7,7 +5,7 @@ void rot2_kernel(
     float *dst,
     const float theta,
     const unsigned int nu, const unsigned int nv, // output size
-    const float sx, const float sy,
+    const float sx, const float sy,               // scale
     const unsigned int nx, const unsigned int ny  // input size
 ) {
     unsigned int iu = blockIdx.x*blockDim.x+threadIdx.x;
@@ -17,11 +15,8 @@ void rot2_kernel(
     }
 
     // move origin to center
-    float u0 = (float)iu - (float)nu/2;
-    float v0 = (float)iv - (float)nv/2;
-
-    // rescale
-    //u0 /= sx; v0 /= sy;
+    float u0 = iu - nu/2.;
+    float v0 = iv - nv/2.;
 
     // rotate
     float x0 =  u0*cosf(theta) + v0*sinf(theta);
@@ -31,8 +26,8 @@ void rot2_kernel(
     x0 /= sx; y0 /= sy;
 
     // move origin to corner
-    float x = x0 + (float)nx/2;
-    float y = y0 + (float)ny/2;
+    float x = x0 + nx/2.;
+    float y = y0 + ny/2.;
 
     // write back
     const unsigned int i = iv*nu + iu;
